refactor: keep task put implementation separate
