--- annofabcli/task/task_creation.py
@@ -1,218 +1,0 @@
-from __future__ import annotations
-
-import logging
-import multiprocessing
-import sys
-import tempfile
-from collections import defaultdict
-from enum import Enum
-from pathlib import Path
-
-import annofabapi
-import pandas
-from annofabapi.models import JobStatus, ProjectJobType
-
-from annofabcli.common.cli import COMMAND_LINE_ERROR_STATUS_CODE, get_json_from_args
-from annofabcli.common.dataclasses import WaitOptions
-
-logger = logging.getLogger(__name__)
-
-
-TASK_THRESHOLD_FOR_JSON = 230
-"""
-タスク作成処理のwebapiに put_taskを使うか initiate_generation_task を使うかのしきい値。
-作成するタスク数が指定した値以下の場合は、put_task webapiを使います。
-しきい値は、実際にwebapiの処理時間を計測して決めました。詳細は以下を参照してください。
-https://github.com/kurusugawa-computer/annofab-cli/pull/738#issuecomment-1077013844
-"""
-
-TASK_RELATION_CSV_COLUMNS = ("task_id", "input_data_id")
-"""タスクと入力データの関係を表すCSVの必須列"""
-
-TaskInputRelation = dict[str, list[str]]
-"""task_idとinput_data_idの構造を表現する型"""
-
-
-class ApiWithCreatingTask(Enum):
-    """タスク作成に使われるWebAPI"""
-
-    PUT_TASK = "put_task"
-    INITIATE_TASKS_GENERATION = "initiate_tasks_generation"
-
-
-def create_task_relation_dict(task_ids: pandas.Series, input_data_ids: pandas.Series) -> TaskInputRelation:
-    """task_idとinput_data_idの列から、辞書を生成する。"""
-
-    result: TaskInputRelation = defaultdict(list)
-    for task_id, input_data_id in zip(task_ids, input_data_ids, strict=False):
-        result[task_id].append(input_data_id)
-    return result
-
-
-def get_task_relation_dict_from_headerless_csv(csv_file: Path) -> TaskInputRelation:
-    """ヘッダ行なしCSVから、keyがtask_id, valueがinput_data_idのlistのdictを生成する。"""
-
-    # `dtype=str`を指定した理由：指定しないと、IDが`001`のときに`1`に変換されてしまうため
-    df = pandas.read_csv(str(csv_file), header=None, usecols=(0, 1), names=TASK_RELATION_CSV_COLUMNS, dtype=str)
-    return create_task_relation_dict(df["task_id"], df["input_data_id"])
-
-
-def get_task_relation_dict_from_header_csv(csv_file: Path) -> TaskInputRelation:
-    """ヘッダ行ありCSVから、keyがtask_id, valueがinput_data_idのlistのdictを生成する。"""
-
-    df = pandas.read_csv(str(csv_file), dtype=str)
-    if "task_id" not in df.columns or "input_data_id" not in df.columns:
-        sys.stderr.write("annofabcli task create: error: CSV形式が不正です。ヘッダ行に 'task_id' と 'input_data_id' を指定してください。\n")
-        sys.exit(COMMAND_LINE_ERROR_STATUS_CODE)
-    return create_task_relation_dict(df["task_id"], df["input_data_id"])
-
-
-def get_task_relation_dict_from_json_args(json_value: str, *, command_name: str) -> TaskInputRelation:
-    """JSON引数からtask_idとinput_data_idの関係を表すdictを取得する。"""
-
-    task_relation_dict = get_json_from_args(json_value)
-    if not isinstance(task_relation_dict, dict):
-        print(f"{command_name}: error: JSON形式が不正です。オブジェクトを指定してください。", file=sys.stderr)  # noqa: T201
-        sys.exit(COMMAND_LINE_ERROR_STATUS_CODE)
-    return task_relation_dict
-
-
-def create_task_relation_csv(task_relation_dict: TaskInputRelation, csv_file: Path) -> None:
-    """task_idとinput_data_idの関係を持つdictから、``initiate_tasks_generation`` APIに渡すCSVを生成する。
-
-    Args:
-        task_relation_dict: keyがtask_id, valueがinput_data_idのlistのdict
-        csv_file: 出力先
-
-    """
-    tmp_list = []
-    for task_id, input_data_id_list in task_relation_dict.items():
-        for input_data_id in input_data_id_list:
-            tmp_list.append({"task_id": task_id, "input_data_id": input_data_id})  # noqa: PERF401
-    df = pandas.DataFrame(tmp_list)
-
-    # webapiの都合上、2列目は空文字でないといけない
-    df["empty"] = ""
-    df = df[["task_id", "empty", "input_data_id"]]
-
-    csv_file.parent.mkdir(exist_ok=True, parents=True)
-    df.to_csv(str(csv_file), index=False, header=None)
-
-
-class TaskCreatingMain:
-    def __init__(
-        self,
-        service: annofabapi.Resource,
-        project_id: str,
-        *,
-        parallelism: int | None,
-        should_wait: bool = False,
-    ) -> None:
-        self.service = service
-        self.project_id = project_id
-        self.parallelism = parallelism
-        self.should_wait = should_wait
-
-    def create_task(self, task_id: str, input_data_id_list: list[str]) -> bool:
-        task = self.service.wrapper.get_task_or_none(self.project_id, task_id)
-        if task is not None:
-            logger.warning(f"タスク'{task_id}'はすでに存在するため、登録をスキップします。")
-            return False
-
-        # タスクを上書きしない理由：タスクを上書きすると、タスクに紐づくアノテーションまで消えてしまう恐れがあるため
-        self.service.api.put_task(self.project_id, task_id, request_body={"input_data_id_list": input_data_id_list})
-        logger.debug(f"タスク'{task_id}'を登録しました。")
-        return True
-
-    def create_task_wrapper(self, tpl: tuple[str, list[str]]) -> bool:
-        task_id, input_data_id_list = tpl
-        try:
-            return self.create_task(task_id, input_data_id_list)
-        except Exception:  # pylint: disable=broad-except
-            logger.warning(f"タスク'{task_id}'の登録に失敗しました。", exc_info=True)
-            return False
-
-    def create_task_list(self, task_relation_dict: TaskInputRelation) -> None:
-        logger.debug("'put_task' WebAPIを用いてタスクを生成します。")
-        success_count = 0
-        if self.parallelism is None:
-            for task_id, input_data_id_list in task_relation_dict.items():
-                try:
-                    result = self.create_task(task_id, input_data_id_list)
-                    if result:
-                        success_count += 1
-                except Exception:  # pylint: disable=broad-except
-                    logger.warning(f"タスク'{task_id}'の登録に失敗しました。", exc_info=True)
-
-        else:
-            with multiprocessing.Pool(self.parallelism) as pool:
-                results = pool.map(self.create_task_wrapper, task_relation_dict.items())
-                success_count = len([e for e in results if e])
-
-        logger.info(f"{success_count} / {len(task_relation_dict)} 件のタスクを登録しました。")
-
-    def create_task_from_csv_file(self, csv_file: Path) -> None:
-        """CSVファイルからタスクを登録する。"""
-
-        logger.debug("'initiate_tasks_generation' WebAPIを用いてタスクを生成します。")
-        content = self.service.wrapper.initiate_tasks_generation_by_csv(self.project_id, csvfile_path=str(csv_file))
-        job = content["job"]
-        logger.info(f"Annofab上でタスク作成処理が開始されました。 :: csv_file='{csv_file}', job_type='{job['job_type']}', job_id='{job['job_id']}'")
-        if self.should_wait:
-            self.wait_for_completion(job["job_id"])
-        else:
-            logger.info(
-                f"以下のコマンドを実行すれば、タスク登録ジョブが終了するまで待ちます。 :: `annofabcli job wait --project_id {self.project_id} --job_type {job['job_type']} --job_id {job['job_id']}`"
-            )
-
-    def generate_task(
-        self,
-        api: ApiWithCreatingTask | None,
-        task_relation_dict: TaskInputRelation,
-    ) -> None:
-        """put_task または initiate_tasks_generation WebAPIを使って、タスクを生成する。"""
-
-        logger.info(f"{len(task_relation_dict)}件のタスクを生成します。")
-        if api is None:
-            if len(task_relation_dict) > TASK_THRESHOLD_FOR_JSON:
-                # `NamedTemporaryFile`を使わない理由: Windowsで`PermissionError`が発生するため
-                with tempfile.TemporaryDirectory() as str_temp_dir:
-                    csv_path = Path(str_temp_dir) / f"{self.project_id}.csv"
-                    create_task_relation_csv(task_relation_dict, csv_path)
-                    self.create_task_from_csv_file(csv_path)
-            else:
-                self.create_task_list(task_relation_dict)
-            return
-
-        if api == ApiWithCreatingTask.PUT_TASK:
-            self.create_task_list(task_relation_dict)
-        elif api == ApiWithCreatingTask.INITIATE_TASKS_GENERATION:
-            # `NamedTemporaryFile`を使わない理由: Windowsで`PermissionError`が発生するため
-            with tempfile.TemporaryDirectory() as str_temp_dir:
-                csv_path = Path(str_temp_dir) / f"{self.project_id}.csv"
-                create_task_relation_csv(task_relation_dict, csv_path)
-                self.create_task_from_csv_file(csv_path)
-
-    def wait_for_completion(self, job_id: str) -> None:
-        """タスク登録ジョブが終了するまで待つ。"""
-
-        wait_options = WaitOptions(interval=30, max_tries=720)
-        max_wait_minute = wait_options.max_tries * wait_options.interval / 60
-        logger.info(f"job_id='{job_id}' :: 最大{max_wait_minute}分間、タスク登録のジョブが終了するまで待ちます。")
-
-        result = self.service.wrapper.wait_until_job_finished(
-            self.project_id,
-            job_type=ProjectJobType.GEN_TASKS,
-            job_id=job_id,
-            job_access_interval=wait_options.interval,
-            max_job_access=wait_options.max_tries,
-        )
-        if result is None:
-            logger.error(f"job_id='{job_id}' :: タスク登録のジョブが存在しません。")
-            return
-        if result == JobStatus.SUCCEEDED:
-            logger.info(f"job_id='{job_id}' :: タスク登録のジョブが成功しました。")
-        elif result == JobStatus.FAILED:
-            logger.error(f"job_id='{job_id}' :: タスク登録のジョブが失敗しました。")
-        elif result == JobStatus.PROGRESS:
-            logger.warning(f"job_id='{job_id}' :: {max_wait_minute}分間待ちましたが、タスク登録のジョブが終了しませんでした。")
